tasks: extract pausable config
